Broken out into separate files for cleaner code structure
