+++ lib/painted_rabbit/field.rb
@@ -1,0 +1,9 @@
+class PaintedRabbit::Field
+  attr_reader :method, :name, :serializer, :options
+  def initialize(method, name, serializer, options = {})
+    @method = method
+    @name = name
+    @serializer = serializer
+    @options = options
+  end
+end
