+++ lib/painted_rabbit/painted_rabbit_error.rb
@@ -1,0 +1,2 @@
+class PaintedRabbit::PaintedRabbitError < StandardError
+end
