+++ lib/painted_rabbit/serializers/public_send_serializer.rb
@@ -1,0 +1,5 @@
+class PaintedRabbit::PublicSendSerializer < PaintedRabbit::Serializer
+  serialize do |field_name, object|
+    object.public_send(field_name)
+  end
+end
